requirements.txt
streamlit
tensorflow
pillow
requests